Switched sample rate back to float. I suppose it's theoretically possible for there to be a non-integral number of samples per second and this is obvious when doing some calculations.

+++ src/test/kotlin/audio/audioInput/AudioInputTests.kt
@@ -34,7 +34,7 @@ class AudioInputTests {
     @BeforeEach
     fun setupHappyPath() {
         every { inputLine.name } returns nextString("name")
-        every { inputLine.sampleRate } returns nextPositiveInt()
+        every { inputLine.sampleRate } returns nextPositiveInt().toFloat()
         every { inputLine.bitDepth } returns nextPositiveInt()
         every { inputLine.channels } returns nextPositiveInt()
